Move data_types out of utils module
