+++ pyhessian/utils/__init__.py
@@ -1,2 +1,1 @@
 from . import toposort
-from . import data_types
