+++ tests/test_encoding.py
@@ -3,7 +3,7 @@ import datetime
 import six
 
 from pyhessian import protocol
-from pyhessian.utils.data_types import long
+from pyhessian.data_types import long
 from .base import HessianTestCase
 
 # Caucho's Hessian 2.0 reference service
